+++ src/hooks/use-link-form.tsx
@@ -63,10 +63,17 @@ function reducer<Field extends FormField>(state: LinkFormState, action: LinkForm
   }
 }
 
-export default function useLinkForm(initialValues: Partial<LinkFormState["values"]> = {}) {
+export type LinkFormOptions = {
+  detectFrontmostLink?: boolean;
+};
+
+export default function useLinkForm(
+  initialValues: Partial<LinkFormState["values"]> = {},
+  { detectFrontmostLink = true }: LinkFormOptions = {}
+) {
   const initialState: LinkFormState = {
     dirty: false,
-    hasUpdatedWithLink: false,
+    hasUpdatedWithLink: !detectFrontmostLink,
     values: {
       description: "",
       favicon: "",