+++ src/types/preferences.tsx
@@ -6,9 +6,15 @@ export type FormActionPreference =
   | "copyUrl"
   | "copyUrlAsMarkdown"
   | "clearCache"
-  | "fetchContent";
+  | "fetchContent"
+  | "saveChanges";
 
-export type DetailActionPreference = FormActionPreference | "showDetails" | "markAsRead" | "deleteFile";
+export type DetailActionPreference =
+  | FormActionPreference
+  | "showDetails"
+  | "editBookmark"
+  | "markAsRead"
+  | "deleteFile";
 
 export interface Preferences {
   vaultPath: string;